==> fair_schaer/4b_eval_runs.ipynb <==
# LongEval
In diesem Notebook findet ihr einen einfachen starter um ein Retrieval System für den LongEval Shared Task aufzusetzen.

Weitere Informationen findet ihr hier:
- [Shared Task Website](https://clef-longeval.github.io/)
- [Overview Paper 2024](http://www.zubiaga.org/publications/files/alkhalifa2024longeval-extended.pdf)
- [Overview Paper 2023](http://www.zubiaga.org/publications/files/alkhalifa2023longeval-overview.pdf)
- [LongEval Test Collection Paper](https://www.semanticscholar.org/reader/f40debce2b7caf35ea0730c27c5330989d20b300)

Alle Datensätze findet ihr unter `datasets/LongEval`. PyTerrier Indexe für jeden Zeitpunkt mit der üblichen Preprocessing Pipeline findet ihr in `datasets/LongEval/index`. In der Datei `datasets/LongEval/metadata.yml` findet ihr Metadaten die euch gegebenenfalls helfen die Sub-Collections zu organisieren.

import yaml
import os
import pandas as pd
import numpy as np

import pyterrier as pt
if not pt.java.started():
    pt.java.init()

from sqlalchemy import create_engine

DATABASE = "longeval-web"
USER = "dis18"
HOST = "db"
PORT = "5432"
PASSWORD = "dis182425"

engine = create_engine(f"postgresql+psycopg2://{USER}:{PASSWORD}@{HOST}:{PORT}/{DATABASE}")

df = pd.read_sql('select * from "Topic" limit 1', con=engine)
sql_query = lambda x: pd.read_sql(x, con=engine)

dataset = "longeval-web"
language = "fr"
sub_collection = "2023-02"

BASE_PATH = "/home/jovyan/work/datasets/LongEval-Web"

with open(BASE_PATH + "/metadata.yml", "r") as yamlfile:
    config = yaml.load(yamlfile, Loader=yaml.FullLoader)

index_path = os.path.join(".", BASE_PATH, f"index/{dataset}-{language}-{sub_collection}-pyterrier")
topics_path = os.path.join(BASE_PATH, "release_2025_p1/French/queries.txt")

q_topics = f"""
select distinct b.queryid qid, b.text_fr query
from "Qrel" a 
join  (
        select  *
        from    "Topic"
      ) b
      on      a.queryid = b.queryid
join (
        select distinct docid
        from   "Document"
        where  sub_collection = '{sub_collection}'
      )c
      on ('doc' || a.docid) = c.docid
where a.sub_collection = '{sub_collection}' 
"""

topics = sql_query(q_topics)
#topics = pd.read_csv(topics_path, sep="\t", names=["qid", "query"])
topics["qid"] = topics["qid"].astype(str)

topics["query"] = topics["query"].str.replace("'", "")
topics["query"] = topics["query"].str.replace("*", "")
topics["query"] = topics["query"].str.replace("/", "")
topics["query"] = topics["query"].str.replace(":", "")
topics["query"] = topics["query"].str.replace("?", "")
topics["query"] = topics["query"].str.replace(")", "")
topics["query"] = topics["query"].str.replace("(", "")
topics["query"] = topics["query"].str.replace("+", "")
spam = ["59769", "6060", "75200", "74351", "67599", "74238", "74207", "75100", "58130"]
topics = topics[~topics["qid"].isin(spam)]

index = pt.IndexFactory.of(index_path)

### Run Erstellen

BM25 = pt.terrier.Retriever(index, wmodel="BM25", verbose=True)

print(">>> Loaded index with", index.getCollectionStatistics().getNumberOfDocuments(), "documents.")

# Run Laden 
Erstellen eines Runs für alle Topics dauert sehr lange. Alternativ könnt ihr auch einen BM25 Baseline Run laden und eure Ansätze als Re-Ranking implementieren.

# Read own run
run = pd.read_csv("run_modelprediction_2023-02.csv")

run_base = pt.io.read_results(f"{BASE_PATH}/runs/{dataset}-{language}-{sub_collection}-BM25.gz")
run_base["docno"] = run_base["docno"].str.strip("doc")  # the indexed documents prefix the docid with `doc`, this needs to be removed

qid_map = {f"{qid}":1 for qid in run["qid"].drop_duplicates().tolist() }
run_base["qid_bool"] = run_base["qid"].map(qid_map)
run_base_small = run_base.dropna(subset=["qid_bool"])

# System Evaluieren

run_base_small["qid"].unique().shape

q_qrels = f"""
select a.queryid qid, a.docid docno, cast(a.relevance as int) label
from "Qrel" a 
join  (
        select  *
        from    "Topic"
      ) b
      on      a.queryid = b.queryid
join (
        select distinct docid
        from   "Document"
        where  sub_collection = '{sub_collection}'
      )c
      on ('doc' || a.docid) = c.docid
where a.sub_collection = '{sub_collection}' 
"""

qrels = sql_query(q_qrels)
#qrels = pt.io.read_qrels(BASE_PATH + f"/release_2025_p1/French/LongEval Train Collection/qrels/{sub_collection}_{language}/qrels_processed.txt")

print(qrels.columns.tolist())

run["docno"] = run["docno"].str.strip("doc")  # the indexed documents prefix the docid with `doc`, this needs to be removed

cols = ["qid", "docno", "rank", "score"]
run_eval = run[cols]
pt.Experiment(
    [run_base_small, run_eval],
    topics,
    qrels,
    eval_metrics=["bpref", "map", "ndcg", "ndcg_cut_10", "P.10"],
    verbose=True
)

print(run_base)

cols = ["qid", "docno", "rank", "score"]
run_eval = run[cols]
df_res = pt.Experiment(
    [run_base_small, run_eval],
    topics,
    qrels,
    eval_metrics=["ndcg"],#["bpref", "map", "ndcg", "ndcg_cut_10", "P.10"],
    verbose=True,
    perquery=True,
)
df_res
