==> fair_schaer/4c_transform_to_trec_run_format.ipynb <==
## Transform `pd.DataFrame`-runs to `run.txt`

import os
import gzip
from pathlib import Path
import pandas as pd

runs = [
    "submit_run_2023-02.csv",
    "submit_run_2023-03.csv",
    "submit_run_2023-04.csv",
    "submit_run_2023-05.csv",
    "submit_run_2023-06.csv",
    "submit_run_2023-07.csv",
    "submit_run_2023-08.csv",
]

l_strip = "submit_run_"
r_strip = ".csv"
all_strip = l_strip + r_strip

BASE_PATH = "submissions/"

trec_cols = ["qid", "placeholder", "docno", "rank", "score", "name"]

for run in runs:
    # Create Directory
    sub_collection = run.strip(all_strip)
    sub_path = BASE_PATH+sub_collection
    sub_run_file = sub_path + "/run.txt"
    Path(sub_path).mkdir(parents=True, exist_ok=True)

    # Transform DataFrame to TREC-EVAL format
    df = pd.read_csv(run)
    df_trec = pd.DataFrame()
    
    df_trec["qid"] = df["qid"]
    df_trec["placeholder"] = "Q0"
    df_trec["docno"] = df["docno"]
    df_trec["rank"] = df["rank"]
    df_trec["score"] = df["score"]    
    df_trec["name"] = sub_collection

    print(df_trec)
    df_trec.to_csv(sub_run_file, index=False, header=False, sep=" ")
    #break
    

## Transform `run.txt` to `run.txt.gz`

The `compress="gzip"` in `df.to_csv()` seems to yield errors with the current format.</br>
A solution has to be made, where the gzip compressions needs to be applied after saving the raw text files.</br>
There is an easy workaround for MAC/Linux Distributions.</br>

*Code*

The code looks for `run.txt` files in the parent- and child directories.</br>
Every file found will then be compressed to its respective `run.txt.gz` file.

`find -name "run.txt" -type f | xargs -I {} gzip {} > "{}.gz"`
